# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateInstalledFirewall:
    properties:
      backup_date:
        type: string
      brand:
        type: string
      created_by_user_id:
        type: integer
      current_version:
        type: string
      customer_lan_on_site:
        type: string
      customer_location_id:
        type: integer
      equipment:
        type: string
      extra:
        type: string
      firewall_ipv4_lan_address:
        type: string
      firewall_ipv6_lan_address:
        type: string
      installed_date:
        type: string
      version_backup:
        type: string
    type: object
  models.CreateInternetProvider:
    properties:
      account_or_pin:
        type: string
      name:
        type: string
      other:
        type: string
      primary_dns:
        type: string
      secondary_dns:
        type: string
      speed:
        type: string
    type: object
  models.CreateNote:
    properties:
      assigned_to_id:
        type: integer
      attachment:
        type: string
      created_by_user_id:
        type: integer
      customer_location_id:
        type: integer
      note:
        type: string
      subject:
        type: string
    type: object
  models.CreateServer:
    properties:
      created_by_user_id:
        type: integer
      customer_location_id:
        type: integer
      days_left:
        type: string
      description:
        type: string
      expiration_date:
        type: string
      hardware_as_a_service:
        type: boolean
      host_name:
        type: string
      idrac:
        type: string
      location:
        type: string
      order_number:
        type: string
      os_platform:
        type: string
      ownership:
        type: string
      power_connect_type:
        type: string
      purchase_date:
        type: string
      service_tag:
        type: string
      type:
        type: string
      warranty:
        type: string
    type: object
...
  models.CreateWanConfig:
    properties:
      comment:
        type: string
      firewall:
        type: string
      gateway:
        type: string
      ip_range:
        type: string
      ip_version:
        type: string
      sub_net_mask:
        type: string
    type: object
...
  models.InstalledFirewall:
    properties:
      backup_date:
        type: string
      brand:
        type: string
      created_at:
        type: string
      created_by_user_id:
        type: integer
      current_version:
        type: string
      customer_lan_on_site:
        type: string
      customer_location_id:
        type: integer
      deleted_at:
        type: string
      deleted_by_user_id:
        type: integer
      equipment:
        type: string
      extra:
        type: string
      firewall_ipv4_lan_address:
        type: string
      firewall_ipv6_lan_address:
        type: string
      installed_date:
        type: string
      installed_firewall_id:
        type: integer
      internet_provider1:
        $ref: '#/definitions/models.InternetProvider'
      internet_provider1_id:
        type: integer
      internet_provider2:
        $ref: '#/definitions/models.InternetProvider'
      internet_provider2_id:
        type: integer
      updated_at:
        type: string
      updated_by_user_id:
        type: integer
      version_backup:
        type: string
    type: object
  models.InternetProvider:
    properties:
      account_or_pin:
        type: string
      name:
        type: string
      other:
        type: string
      primary_dns:
        type: string
      secondary_dns:
        type: string
      speed:
        type: string
      wan_config_ipv4:
        $ref: '#/definitions/models.WanConfig'
      wan_config_ipv4_id:
        type: integer
      wan_config_ipv6:
        $ref: '#/definitions/models.WanConfig'
      wan_config_ipv6_id:
        type: integer
    type: object
...
  models.Note:
    properties:
      assigned_to_id:
        type: integer
      attachment:
        type: string
      created_at:
        type: string
      created_by_user_id:
        type: integer
      customer_location_id:
        type: integer
      deleted_at:
        type: string
      deleted_by_user_id:
        type: integer
      note:
        type: string
      note_id:
        type: integer
      subject:
        type: string
      updated_at:
        type: string
      updated_by_user_id:
        type: integer
    type: object
  models.Server:
    properties:
      created_at:
        type: string
      created_by_user_id:
        type: integer
      customer_location_id:
        type: integer
      days_left:
        type: string
      deleted_at:
        type: string
      deleted_by_user_id:
        type: integer
      description:
        type: string
      expiration_date:
        type: string
      expression_service_code:
        type: integer
      hardware_as_a_service:
        type: boolean
      host_name:
        type: string
      idrac:
        type: string
      location:
        type: string
      order_number:
        type: string
      os_platform:
        type: string
      ownership:
        type: string
      power_connect_type:
        type: string
      purchase_date:
        type: string
      server_id:
        type: integer
      service_tag:
        type: string
      type:
        type: string
      updated_at:
        type: string
      updated_by_user_id:
        type: integer
      warranty:
        type: string
    type: object
...
  models.UpdateInstalledFirewall:
    properties:
      backup_date:
        type: string
      brand:
        type: string
      current_version:
        type: string
      customer_lan_on_site:
        type: string
      customer_location_id:
        type: integer
      equipment:
        type: string
      extra:
        type: string
      firewall_ipv4_lan_address:
        type: string
      firewall_ipv6_lan_address:
        type: string
      installed_date:
        type: string
      updated_by_user_id:
        type: integer
      version_backup:
        type: string
    type: object
  models.UpdateInternetProvider:
    properties:
      account_or_pin:
        type: string
      name:
        type: string
      other:
        type: string
      primary_dns:
        type: string
      secondary_dns:
        type: string
      speed:
        type: string
    type: object
  models.UpdateNote:
    properties:
      assigned_to_id:
        type: integer
      attachment:
        type: string
      customer_location_id:
        type: integer
      note:
        type: string
      subject:
        type: string
      updated_by_user_id:
        type: integer
    type: object
  models.UpdateServer:
    properties:
      customer_location_id:
        type: integer
      days_left:
        type: string
      description:
        type: string
      expiration_date:
        type: string
      expression_service_code:
        type: integer
      hardware_as_a_service:
        type: boolean
      host_name:
        type: string
      idrac:
        type: string
      location:
        type: string
      order_number:
        type: string
      os_platform:
        type: string
      ownership:
        type: string
      power_connect_type:
        type: string
      purchase_date:
        type: string
      service_tag:
        type: string
      type:
        type: string
      updated_by_user_id:
        type: integer
      warranty:
        type: string
    type: object
...
  models.UpdateWanConfig:
    properties:
      comment:
        type: string
      firewall:
        type: string
      gateway:
        type: string
      ip_range:
        type: string
      ip_version:
        type: string
      sub_net_mask:
        type: string
    type: object
...
  models.WanConfig:
    properties:
      comment:
        type: string
      firewall:
        type: string
      gateway:
        type: string
      ip_range:
        type: string
      ip_version:
        type: string
      sub_net_mask:
        type: string
      wan_config_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server .
  termsOfService: http://swagger.io/terms/
  title: Network Management Solutions API
  version: "1.0"
paths:
...
  /installed-firewalls:
    delete:
      description: Delete Installed Firewall
      parameters:
      - description: Installed Firewall ID
        in: query
        name: installed_firewall_id
        type: string
      - description: USER ID
        in: header
        name: USER_ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InstalledFirewall'
      summary: Delete Installed Firewall
      tags:
      - Installed Firewall
    get:
      description: Retrieve  Installed Firewall
      parameters:
      - description: Domain
        in: query
        name: version_backup
        type: string
      - description: Email Domain Id
        in: query
        name: firewall_wan1_ipv4
        type: string
      - description: Customer Location
        in: query
        name: backup_date
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Order By
        in: query
        name: order_by
        type: string
      - description: Select Column
        in: query
        name: select_column
        type: string
      - description: Append Select
        in: query
        name: append_select
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InstalledFirewall'
      summary: Retrieve  Installed Firewall
      tags:
      - Installed Firewall
    patch:
      consumes:
      - application/json
      description: Update Installed Firewall
      parameters:
      - description: Installed Firewall
        in: body
        name: InstalledFirewall
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/models.UpdateInstalledFirewall'
          - properties:
              internet_provider1:
                allOf:
                - $ref: '#/definitions/models.UpdateInternetProvider'
                - properties:
                    wan_config_ipv4:
                      $ref: '#/definitions/models.UpdateWanConfig'
                    wan_config_ipv6:
                      $ref: '#/definitions/models.UpdateWanConfig'
                  type: object
              internet_provider2:
                allOf:
                - $ref: '#/definitions/models.UpdateInternetProvider'
                - properties:
                    wan_config_ipv4:
                      $ref: '#/definitions/models.UpdateWanConfig'
                    wan_config_ipv6:
                      $ref: '#/definitions/models.UpdateWanConfig'
                  type: object
            type: object
      - description: Installed Firewall ID
        in: query
        name: installed_firewall_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InstalledFirewall'
      summary: Update Installed Firewall
      tags:
      - Installed Firewall
    post:
      consumes:
      - application/json
      description: Add    Installed Firewall
      parameters:
      - description: InstalledFirewall Data
        in: body
        name: InstalledFirewall
        required: true
        schema:
          allOf:
          - $ref: '#/definitions/models.CreateInstalledFirewall'
          - properties:
              internet_provider1:
                allOf:
                - $ref: '#/definitions/models.CreateInternetProvider'
                - properties:
                    wan_config_ipv4:
                      $ref: '#/definitions/models.CreateWanConfig'
                    wan_config_ipv6:
                      $ref: '#/definitions/models.CreateWanConfig'
                  type: object
              internet_provider2:
                allOf:
                - $ref: '#/definitions/models.CreateInternetProvider'
                - properties:
                    wan_config_ipv4:
                      $ref: '#/definitions/models.CreateWanConfig'
                    wan_config_ipv6:
                      $ref: '#/definitions/models.CreateWanConfig'
                  type: object
            type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.InstalledFirewall'
      summary: Create Installed Firewall
      tags:
      - Installed Firewall
...
  /notes:
    delete:
      description: Delete Notes
      parameters:
      - description: Note ID
        in: query
        name: note_id
        type: string
      - description: USER ID
        in: header
        name: USER_ID
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Note'
      summary: Delete Notes
      tags:
      - Notes
    get:
      description: Retrieve  Notes
      parameters:
      - description: Note ID
        in: query
        name: note_id
        type: string
      - description: Subject
        in: query
        name: subject
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Order By
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Note'
      summary: Retrieve  Notes
      tags:
      - Notes
    patch:
      consumes:
      - application/json
      description: Update Notes
      parameters:
      - description: Note
        in: body
        name: Note
        required: true
        schema:
          $ref: '#/definitions/models.UpdateNote'
      - description: Note ID
        in: query
        name: note_id
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Note'
      summary: Update Notes
      tags:
      - Notes
    post:
      consumes:
      - application/json
      description: Add    Notes
      parameters:
      - description: Note Data
        in: body
        name: Note
        required: true
        schema:
          $ref: '#/definitions/models.CreateNote'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Note'
      summary: Create Notes
      tags:
      - Notes
  /servers:
    delete:
      description: Delete Server
      parameters:
      - description: Server ID
        in: query
        name: server_id
        type: string
      - description: USER ID
        in: header
        name: USER_ID
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Server'
      summary: Delete Server
      tags:
      - Servers
    get:
      description: Retrieve  Server
      parameters:
      - description: Server ID
        in: query
        name: server_id
        type: string
      - description: Service Tag
        in: query
        name: service_tag
        type: string
      - description: equipment type
        in: query
        name: equipment_type
        type: string
      - description: type
        in: query
        name: type
        type: string
      - description: expiration date
        in: query
        name: expiration_date
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Order By
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Note'
      summary: Retrieve  Server
      tags:
      - Servers
    patch:
      consumes:
      - application/json
      description: Update Server
      parameters:
      - description: Server
        in: body
        name: Server
        required: true
        schema:
          $ref: '#/definitions/models.UpdateServer'
      - description: Server ID
        in: query
        name: server_id
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Server'
      summary: Update Server
      tags:
      - Servers
    post:
      consumes:
      - application/json
      description: Add    Server
      parameters:
      - description: Server Data
        in: body
        name: Server
        required: true
        schema:
          $ref: '#/definitions/models.CreateServer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Server'
      summary: Create Server
      tags:
      - Servers
...
schemes:
- http
swagger: "2.0"